requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.5.0
opencv-python>=4.8.0
scikit-learn>=1.3.0

# Face recognition and embeddings
facenet-pytorch>=2.5.3
timm>=0.9.0

# CLIP
transformers>=4.30.0
clip-by-openai>=1.0

# FAISS for similarity search
faiss-cpu>=1.7.4
# For GPU support, use: faiss-gpu>=1.7.4

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
grad-cam>=1.4.8

# Data handling
tqdm>=4.65.0
h5py>=3.9.0

# Web framework (for future backend)
flask>=2.3.0
flask-cors>=4.0.0

# Utilities
pyyaml>=6.0
requests>=2.31.0
